Move ConnectionEstablisher to /app directory
