+++ src/renderer/store/modules/connection.js
@@ -30,7 +30,7 @@ import ConsumerLocationDTO from '../../../libraries/mysterium-tequilapi/dto/cons
 import type { BugReporter } from '../../../app/bug-reporting/interface'
 import logger from '../../../app/logger'
 import TequilapiError from '../../../libraries/mysterium-tequilapi/tequilapi-error'
-import type { ConnectionActions, ConnectionEstablisher } from '../connection-establisher'
+import type { ConnectionActions, ConnectionEstablisher } from '../../../app/connection/connection-establisher'
 
 type ConnectionStore = {
   ip: ?string,
