+++ test/unit/app/connection/connection-establisher.spec.js
@@ -20,10 +20,10 @@
 import { beforeEach, describe, expect, it } from '../../../helpers/dependencies'
 import ConnectionStatusEnum from '../../../../src/libraries/mysterium-tequilapi/dto/connection-status-enum'
 import ConnectionRequestDTO from '../../../../src/libraries/mysterium-tequilapi/dto/connection-request'
-import TequilapiConnectionEstablisher from '../../../../src/renderer/store/connection-establisher'
+import TequilapiConnectionEstablisher from '../../../../src/app/connection/connection-establisher'
 import MockEventSender from '../../../helpers/statistics/mock-event-sender'
 import BugReporterMock from '../../../helpers/bug-reporter-mock'
-import type { ConnectionActions } from '../../../../src/renderer/store/connection-establisher'
+import type { ConnectionActions } from '../../../../src/app/connection/connection-establisher'
 import ConsumerLocationDTO from '../../../../src/libraries/mysterium-tequilapi/dto/consumer-location'
 import type { ConnectionStore } from '../../../../src/renderer/store/modules/connection'
 import type { ConnectionStatus } from '../../../../src/libraries/mysterium-tequilapi/dto/connection-status-enum'
